admin can see orders

+++ backend/routes/router.js
@@ -4,15 +4,16 @@ const router = express.Router();
 const schemas = require("../models/schemas");
 
 const verifyJWT = (req, res, next) => {
-  const token = req.headers["x-access-token"];
+  const authHeader = req.headers["x-access-token"];
+  const token = authHeader && authHeader.split(" ")[1];
   if (!token) {
     res.send("Nicht Angemeldet");
   } else {
-    jwt.verify(token, process.env.SECRET, (err, decoded) => {
+    jwt.verify(token, process.env.ACCESS_TOKEN_SECRET, (err, user) => {
       if (err) {
         res.json({ auth: false, message: "Zugriff nicht erlaubt" });
       } else {
-        req.userID = decoded.id;
+        req.user = user;
         next();
       }
     });
@@ -72,6 +73,35 @@ router.get("/ingredients", async (req, res) => {
   }
 });
 
+router.get("/admin/orders", verifyJWT, async (req, res) => {
+  if (req.user.role === "admin") {
+    try {
+      const adminOrder = schemas.AdminOrder;
+      const allOrder = await adminOrder.find({}).exec();
+      res.send({ auth: true, allOrder: allOrder });
+    } catch (err) {
+      res.status(500).json(err);
+    }
+  } else {
+    res.json({ auth: false, message: "Zugriff nicht erlaubt" });
+  }
+});
+router.delete("/admin/orders/:id", verifyJWT, async (req, res) => {
+  if (req.user.role === "admin") {
+    const id = req.params.id;
+    try {
+      const adminOrder = schemas.AdminOrder;
+      await adminOrder.deleteOne({ _id: id });
+      const answer = "Sucessfully Deleted _id: " + id;
+      res.send(answer);
+    } catch (err) {
+      res.status(500).json(err);
+    }
+  } else {
+    res.json({ auth: false, message: "Zugriff nicht erlaubt" });
+  }
+});
+
 //Authentifizierung einbauen
 router.post("/orders", async (req, res) => {
   const {
@@ -102,6 +132,10 @@ router.post("/orders", async (req, res) => {
     ordered_items: ordered_items,
     entryDate: entryDate,
   };
+  //save for admin
+  const adminNewOrder = new schemas.AdminOrder(orederData);
+  await adminNewOrder.save();
+  //save for employee
   const newOrder = new schemas.Order(orederData);
   //always use await when interact with url|database
   const saveOrder = await newOrder.save();
@@ -112,40 +146,66 @@ router.post("/orders", async (req, res) => {
   }
   res.end();
 });
+
 //Authentifizierung einbauen
 router.get("/orders", verifyJWT, async (req, res) => {
-  try {
-    const order = schemas.Order;
-    const allOrder = await order.find({}).exec();
-    res.send({ auth: true, allOrder: allOrder });
-  } catch (err) {
-    res.status(500).json(err);
+  if (req.user.role === "employee" || req.user.role === "admin") {
+    try {
+      const order = schemas.Order;
+      const allOrder = await order.find({}).exec();
+      res.send({ role: req.user.role, auth: true, allOrder: allOrder });
+    } catch (err) {
+      res.status(500).json(err);
+    }
+  } else {
+    res.json({ auth: false, message: "Zugriff nicht erlaubt" });
   }
 });
 
 router.delete("/orders/:id", verifyJWT, async (req, res) => {
-  const id = req.params.id;
-  try {
-    const order = schemas.Order;
-    await order.deleteOne({ _id: id });
-    const answer = "Sucessfully Deleted _id: " + id;
-    res.send(answer);
-  } catch (err) {
-    res.status(500).json(err);
+  if (req.user.role === "employee" || req.user.role === "admin") {
+    const id = req.params.id;
+    try {
+      const order = schemas.Order;
+      await order.deleteOne({ _id: id });
+      const answer = "Sucessfully Deleted _id: " + id;
+      res.send(answer);
+    } catch (err) {
+      res.status(500).json(err);
+    }
+  } else {
+    res.json({ auth: false, message: "Zugriff nicht erlaubt" });
   }
 });
+
 router.post("/login", async (req, res) => {
   const { username, password } = req.body;
   if (
+    username === process.env.EMPLOYEE_USERNAME &&
+    password === process.env.EMPLOYEE_PASSWORD
+  ) {
+    const token = jwt.sign(
+      { username, role: "employee" },
+      process.env.ACCESS_TOKEN_SECRET,
+      {
+        expiresIn: 1200,
+      }
+    );
+
+    res.json({ role: "employee", auth: true, token: token });
+  } else if (
     username === process.env.ADMIN_USERNAME &&
     password === process.env.ADMIN_PASSWORD
   ) {
-    const id = process.env.ADMIN_ID;
-    const token = jwt.sign({ id }, process.env.SECRET, {
-      expiresIn: 300 * 96,
-    });
+    const token = jwt.sign(
+      { username, role: "admin" },
+      process.env.ACCESS_TOKEN_SECRET,
+      {
+        expiresIn: 1200,
+      }
+    );
 
-    res.json({ auth: true, token: token });
+    res.json({ role: "admin", auth: true, token: token });
   } else {
     res.json({ auth: false, message: "Wrong Credentials!" });
   }
